+++ src/app.tsx
@@ -10,9 +10,9 @@ import LoadingScreen from './loading-screen.tsx';
 
 
 function App(): JSX.Element {
-  const isOffersDataLoading = useAppSelector((state) => state.isOffersDataLoading);
+  const isOffersDataLoading = useAppSelector((state) => state.offers.isOffersDataLoading);
   const offers = useAppSelector((state) => state.offers);
-  const cityOffers = useAppSelector((state) => state.cityOffers);
+  const cityOffers = useAppSelector((state) => state.offers.cityOffers);
   if (isOffersDataLoading) {
     return (
       <LoadingScreen />
